requirements.txt
fastapi
uvicorn[standard]
streamlit
python-dotenv
pydantic

gitpython

pinecone
langchain
langgraph

sentence-transformers
torch
